+++ application/views/template.php
@@ -42,8 +42,30 @@
     </nav>
 
     <div class="container">
-        <?php $this->load->view($main_view);
-         ?>
+        <!-- NOTIFICATION -->
+        <?php 
+            $success = $this->session->flashdata('success');
+            $failed = $this->session->flashdata('failed');
+
+            if (!empty($failed)) {
+                echo '
+                    <div class="alert alert-danger alert-dismissible" role="alert">
+                        <button type="button" class="close" data-dismiss="alert" aria-label="Close"><span aria-hidden="true">&times;</span></button>
+                        <i class="fa fa-times-circle"></i> '.$failed.'
+                    </div>
+                ';
+            }
+
+            if (!empty($success)) {
+                echo '
+                    <div class="alert alert-success alert-dismissible" role="alert">
+                        <button type="button" class="close" data-dismiss="alert" aria-label="Close"><span aria-hidden="true">&times;</span></button>
+                        <i class="fa fa-check-circle"></i> '.$success.'
+                    </div>
+                ';
+            }
+        ?>
+        <?php $this->load->view($main_view); ?>
     </div>
 
 </body>